Extract goaltracker as separated microservice, add crud for goals
