--- PopugJira.Application/Dto/GoalCreateDto.cs
@@ -1,7 +1,0 @@
-namespace PopugJira.Application.Dto
-{
-    public record GoalCreateDto
-    {
-        public string Description { get; init; }
-    }
-}
