--- PopugJira.Application/PopugJira.Application.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\PopugJira.DataAccessLayer.Contract\PopugJira.DataAccessLayer.Contract.csproj" />
-      <ProjectReference Include="..\PopugJira.Domain\PopugJira.Domain.csproj" />
-    </ItemGroup>
-
-</Project>
