--- PopugJira.Application/Services/GoalTrackerService.cs
@@ -1,45 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using PopugJira.Application.Dto;
-using PopugJira.DataAccessLayer.Contract;
-using PopugJira.Domain;
-
-namespace PopugJira.Application.Services
-{
-    public class GoalTrackerService
-    {
-        private readonly IGoalsDataContext goalsDataContext;
-        private readonly IGoalStatesDataContext goalStatesDataContext;
-
-        public GoalTrackerService(IGoalsDataContext goalsDataContext, 
-                                  IGoalStatesDataContext goalStatesDataContext)
-        {
-            this.goalsDataContext = goalsDataContext;
-            this.goalStatesDataContext = goalStatesDataContext;
-        }
-        
-        public async Task<Goal> GetGoal(int id)
-        {
-            return await goalsDataContext.Get(id);
-        }
-        
-        public async Task CreateGoal(GoalCreateDto createDto)
-        {
-            var openState = await goalStatesDataContext.GetOpenState();
-            var goal = new Goal(null, createDto.Description, openState);
-            await goalsDataContext.Create(goal);
-        }
-
-        public async Task OpenGoal(int id)
-        {
-            var openState = await goalStatesDataContext.GetOpenState();
-            await goalsDataContext.SetState(id, openState.Id);
-        }
-
-        public async Task CloseGoal(int id)
-        {
-            var closedState = await goalStatesDataContext.GetClosedState();
-            await goalsDataContext.SetState(id, closedState.Id);
-        }
-    }
-}
