--- PopugJira.DataAccessLayer.Contract/IGoalStatesDataContext.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-using PopugJira.Domain;
-
-namespace PopugJira.DataAccessLayer.Contract
-{
-    public interface IGoalStatesDataContext
-    {
-        Task<GoalState> GetOpenState();
-        Task<GoalState> GetClosedState();
-    }
-}
