--- PopugJira.DataAccessLayer.Contract/IGoalsDataContext.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using PopugJira.Domain;
-
-namespace PopugJira.DataAccessLayer.Contract
-{
-    public interface IGoalsDataContext
-    {
-        Task<int> Create(Goal goal);
-        Task<Goal> Get(int id);
-        Task Delete(int id);
-
-        Task SetState(int goalId, int goalStateId);
-    }
-}
