--- PopugJira.DataAccessLayer.Contract/PopugJira.DataAccessLayer.Contract.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\PopugJira.Domain\PopugJira.Domain.csproj" />
-    </ItemGroup>
-
-</Project>
