--- PopugJira.DataAccessLayer/PopugJira.DataAccessLayer.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\PopugJira.DataAccessLayer.Contract\PopugJira.DataAccessLayer.Contract.csproj" />
-      <ProjectReference Include="..\PopugJira.Domain\PopugJira.Domain.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="FluentMigrator" Version="3.2.15" />
-      <PackageReference Include="FluentMigrator.Runner" Version="3.2.15" />
-      <PackageReference Include="FluentMigrator.Runner.SQLite" Version="3.2.15" />
-      <PackageReference Include="linq2db.AspNet" Version="3.2.3" />
-      <PackageReference Include="linq2db.SQLite" Version="3.2.3" />
-    </ItemGroup>
-
-</Project>
