--- PopugJira.Domain/Definitions/SystemGoalState.cs
@@ -1,9 +1,0 @@
-namespace PopugJira.Domain.Definitions
-{
-    public enum SystemGoalState
-    {
-        Custom,
-        Open,
-        Closed
-    }
-}
