+++ PopugJira.GoalTracker.Application/Dto/GoalCreateDto.cs
@@ -1,0 +1,7 @@
+namespace PopugJira.GoalTracker.Application.Dto
+{
+    public record GoalCreateDto
+    {
+        public string Description { get; set; }
+    }
+}
