+++ PopugJira.GoalTracker.Application/Dto/GoalUpdateDto.cs
@@ -1,0 +1,7 @@
+namespace PopugJira.GoalTracker.Application.Dto
+{
+    public record GoalUpdateDto : GoalCreateDto
+    {
+        public int Id { get; set; }
+    }
+}
