+++ PopugJira.GoalTracker.Application/PopugJira.GoalTracker.Application.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\PopugJira.GoalTracker.DataAccessLayer.Contract\PopugJira.GoalTracker.DataAccessLayer.Contract.csproj" />
+      <ProjectReference Include="..\PopugJira.GoalTracker.Domain\PopugJira.GoalTracker.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
