+++ PopugJira.GoalTracker.DataAccessLayer.Contract/IGoalStatesDataContext.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using PopugJira.GoalTracker.Domain;
+
+namespace PopugJira.GoalTracker.DataAccessLayer.Contract
+{
+    public interface IGoalStatesDataContext
+    {
+        Task<GoalState> GetOpenState();
+        Task<GoalState> GetClosedState();
+    }
+}
