+++ PopugJira.GoalTracker.DataAccessLayer.Contract/PopugJira.GoalTracker.DataAccessLayer.Contract.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\PopugJira.GoalTracker.Domain\PopugJira.GoalTracker.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
