+++ PopugJira.GoalTracker.DataAccessLayer/PopugJira.GoalTracker.DataAccessLayer.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\PopugJira.GoalTracker.DataAccessLayer.Contract\PopugJira.GoalTracker.DataAccessLayer.Contract.csproj" />
+      <ProjectReference Include="..\PopugJira.GoalTracker.Domain\PopugJira.GoalTracker.Domain.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="FluentMigrator" Version="3.2.15" />
+      <PackageReference Include="FluentMigrator.Runner" Version="3.2.15" />
+      <PackageReference Include="FluentMigrator.Runner.SQLite" Version="3.2.15" />
+      <PackageReference Include="linq2db.AspNet" Version="3.2.3" />
+      <PackageReference Include="linq2db.SQLite" Version="3.2.3" />
+    </ItemGroup>
+
+</Project>
