+++ PopugJira.GoalTracker.Domain/Definitions/SystemGoalState.cs
@@ -1,0 +1,9 @@
+namespace PopugJira.GoalTracker.Domain.Definitions
+{
+    public enum SystemGoalState
+    {
+        Custom,
+        Open,
+        Closed
+    }
+}
