+++ PopugJira.GoalTracker.Tests/TestBase.cs
@@ -1,0 +1,12 @@
+using NUnit.Framework;
+
+namespace PopugJira.GoalTracker.Tests
+{
+    public class TestBase
+    {
+        [SetUp]
+        protected virtual void Setup()
+        {
+        }
+    }
+}
