+++ PopugJira.GoalTracker/PopugJira.GoalTracker.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\PopugJira.GoalTracker.Application\PopugJira.GoalTracker.Application.csproj" />
+      <ProjectReference Include="..\PopugJira.GoalTracker.DataAccessLayer\PopugJira.GoalTracker.DataAccessLayer.csproj" />
+    </ItemGroup>
+
+</Project>
