+++ PopugJira.GoalTracker/appsettings.Development.json
@@ -5,6 +5,5 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
-  },
-  "AllowedHosts": "*"
+  }
 }
