+++ PopugJira.GoalTracker/appsettings.json
@@ -6,5 +6,8 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "SQLite": "Data Source=popugjira.sqlite"
+  }
 }
