--- PopugJira.Tests/TestBase.cs
@@ -1,12 +1,0 @@
-using NUnit.Framework;
-
-namespace PopugJira.Tests
-{
-    public class TestBase
-    {
-        [SetUp]
-        protected virtual void Setup()
-        {
-        }
-    }
-}
