+++ PopugJira/Pages/Goals.razor
@@ -1,0 +1,6 @@
+@page "/goals"
+<h3>Goals here</h3>
+
+@code {
+    
+}
