+++ PopugJira/Shared/NavMenu.razor
@@ -11,6 +11,9 @@
             <NavLink class="nav-link" href="" Match="NavLinkMatch.All">
                 <span class="oi oi-home" aria-hidden="true"></span> Home
             </NavLink>
+            <NavLink class="nav-link" href="goals" Match="NavLinkMatch.All">
+                <span class="oi oi-list" aria-hidden="true"></span> Goals
+            </NavLink>
         </li>
     </ul>
 </div>
